ACM-33187: address impl review comments
